# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Deni",
        "description": "Descripción de la API de Ejemplo",
        "termsOfService": "https://www.ejemplo.com/terms",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "license": {
            "name": "Licencia de Ejemplo",
            "url": "https://www.ejemplo.com/licencia"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/rescates": {
            "get": {
                "tags": [
                    "Rescates"
                ],
                "summary": "Obtener todos los rescates",
                "description": "Recupera todos los rescates de la base de datos.",
                "operationId": "indexRescates",
                "responses": {
                    "200": {
                        "description": "Lista de rescates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/rescate"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rescates"
                ],
                "summary": "Registrar un nuevo rescate",
                "description": "Crear un nuevo rescate con la información proporcionada.",
                "operationId": "storeResc",
                "requestBody": {
                    "description": "Datos del rescate",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "direccion",
                                    "estado",
                                    "fecha_rescate",
                                    "id_animal"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": 1
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Dirección del rescate"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "En curso"
                                    },
                                    "fecha_rescate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-11-19"
                                    },
                                    "informacion_adicional": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "id_animal": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rescate registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rescate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rescates/store2": {
            "post": {
                "tags": [
                    "Rescates"
                ],
                "summary": "Registrar un nuevo rescate con información detallada del animal",
                "description": "Crear un nuevo rescate con información detallada del animal proporcionada.",
                "operationId": "store2resc",
                "requestBody": {
                    "description": "Datos del rescate con información detallada del animal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "direccion",
                                    "estado",
                                    "fecha_rescate",
                                    "id_animal"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": 1
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Dirección del rescate"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "En curso"
                                    },
                                    "fecha_rescate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-11-19"
                                    },
                                    "informacion_adicional": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "id_animal": {
                                        "required": [
                                            "nombre",
                                            "especie"
                                        ],
                                        "properties": {
                                            "nombre": {
                                                "type": "string",
                                                "format": "string",
                                                "example": "Nombre del animal"
                                            },
                                            "especie": {
                                                "type": "string",
                                                "format": "string",
                                                "example": "Especie del animal"
                                            },
                                            "raza": {
                                                "type": "string",
                                                "format": "string",
                                                "example": "Raza del animal"
                                            },
                                            "edad": {
                                                "type": "integer",
                                                "format": "int",
                                                "example": 3
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "format": "string",
                                                "example": "Descripción del animal"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rescate registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rescate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rescates/{id}": {
            "get": {
                "tags": [
                    "Rescates"
                ],
                "summary": "Obtener información de un rescate",
                "description": "Obtiene información de un rescate por su ID.",
                "operationId": "showResc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rescate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del rescate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rescate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rescate no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Rescate no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Rescates"
                ],
                "summary": "Actualizar un rescate",
                "description": "Actualizar la información de un rescate por su ID.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rescate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos actualizados del rescate",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "direccion",
                                    "estado",
                                    "fecha_rescate",
                                    "id_animal"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": 1
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nueva dirección del rescate"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Completado"
                                    },
                                    "fecha_rescate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-11-20"
                                    },
                                    "informacion_adicional": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nueva información adicional"
                                    },
                                    "id_animal": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rescate actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rescate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rescate no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Rescate no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "rescate": {
                "title": "rescate",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "id_usuario": {
                        "type": "integer"
                    },
                    "direccion": {
                        "type": "string"
                    },
                    "estado": {
                        "type": "string"
                    },
                    "fecha_rescate": {
                        "type": "string",
                        "format": "date"
                    },
                    "informacion_adicional": {
                        "type": "string"
                    },
                    "id_animal": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}